Create server argument
Run imagej with argument "--server" to run the imagej-server in the main
thread. Optional arguments "-no-headless" and "-no-blocking" are also
available.
